refactor: move source_verifier to utils
